+++ common.h
@@ -4,7 +4,7 @@
 
 
 ssize_t getFileSize(int fd);
-int readIntoBuffer(int fd, void* buff,  size_t size);
-int printBuffer(const void* buff, unsigned size);
+int readIntoBuffer(int fd, void* buff, size_t size);
+int printBuffer(const void* buff, size_t size);
 
 #endif
